web WIP

+++ streamed-payment/contracts/Sheremetyevo.sol
@@ -84,7 +84,7 @@ contract Sheremetyevo {
     function withdraw(address user, address to, int256 amount) public only_for_server {
         require(amount > 0, "Amount must be greater than zero");
         require(_balances.get(user) >= amount, "Balance is not enough");
-        payable(to).send(amount);
+        payable(to).send(uint256(amount));
         setBalance(user,  _balances.get(user) - amount);
     }
